Updates updates
Fixed a potential performance issue that would occur where the random number selected (which would be used to pick an image for assessment as a potential wallpaper) would be predicted as disliked, then that same image would run again by the pure chance that its number could be drawn again. It was possible (not likely) that it could get stuck in that loop forever. Now past images that are not approved get marked and will no longer be assessed. 

... It is possible though that all available options could end up being marked leaving an empty list of options. That should result in random.choice throwing an exception... Again, this isn't likely but this is likely the next thing I'll address.